fix(as-5914): send token to user on get of enter code page
